requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.0.0
tqdm>=4.60.0
opencv-python>=4.5.0

# Machine learning
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers

# TensorBoard logging
tensorboard>=2.8.0

# Optional: for advanced attention visualization
# x-transformers>=1.4.0

# Development dependencies (optional)
# jupyter>=1.0.0
# ipykernel>=6.0.0
